requirements.txt
anthropic>=0.40.0
pinecone>=8.0.0
google-genai>=1.0.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
flask>=3.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
